+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Version 0.2 adds the ability to persist the Tailscale IP address between runs.
 - Configuration order has been flipped so that
   the last configuration file passed on the command-line
   has the highest precedence.
+- Upgraded `tsnet` to 1.38.1.
 
 ## [0.1.1][] - 2022-09-19